finetune/requirements.txt
accelerate==1.0.1
addict==2.4.0
annotated-types==0.7.0
anyio==4.2.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
art==6.4
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
blobfile==2.1.1
bottle==0.12.15
braceexpand==0.1.6
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.8
colorama==0.4.3
comm==0.2.1
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
DALL-E==0.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
einops==0.7.0
entrypoints==0.3
eval_type_backport==0.2.2
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.6.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.13.3
fqdn==1.5.1
fsspec==2024.3.1
ftfy==6.2.3
future==0.18.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
Glances==3.1.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.3
httplib2==0.14.0
httpx==0.26.0
huggingface-hub==0.22.2
idna==2.8
imageio==2.35.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
influxdb==5.2.0
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.12.3
ipython_genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.4
lmdb==1.4.1
lockfile==0.12.2
louis==3.12.0
lxml==4.9.4
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.1.2
matplotlib-inline==0.1.6
mistune==3.0.2
mlxtend==0.23.4
mmcv-full==1.3.9
monotonic==1.5
mpmath==1.3.0
munkres==1.1.4
mypy==1.9.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
open_clip_torch==2.32.0
opencv-python==4.9.0.80
openslide-python==1.4.1
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
prettytable==3.11.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.29.4
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==2.0.0
pyasn1==0.4.2
pycairo==1.16.2
pycm==4.3
pycparser==2.21
pycryptodomex==3.20.0
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pysmi==0.3.2
pysnmp==4.4.6
PySnooper==1.2.1
PySocks==1.7.1
pystache==0.5.4
pytest==8.1.1
pytest-runner==6.0.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.7
pytz==2025.2
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2024.11.6
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.2
sentry-sdk==2.25.1
setproctitle==1.3.5
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.12
system-service==0.3
systemd-python==234
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.4.0
tifffile==2023.7.10
tiktoken==0.7.0
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
topk @ git+https://github.com/oval-group/smooth-topk.git@12c1645f187e2fa0c05f47bf1fe48864d4bd2707
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
TorchSnooper==0.8
torchvision==0.19.1+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
transformers-stream-generator==0.0.4
triton==3.0.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.13.1
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-tweak-tool==0.0.7
uri-template==1.3.0
urllib3==2.2.3
usb-creator==0.3.7
wadllib==1.3.3
wandb==0.19.9
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
xkit==0.0.0
yapf==0.40.2
zipp==3.17.0
